update to ae2wtlibs new JiJ api
