--- src/main/java/com/glodblock/github/extendedae/client/gui/GuiWirelessExPAT.java
@@ -1,35 +1,0 @@
-package com.glodblock.github.extendedae.client.gui;
-
-import appeng.api.upgrades.Upgrades;
-import appeng.client.gui.style.ScreenStyle;
-import appeng.client.gui.widgets.ToolboxPanel;
-import appeng.client.gui.widgets.UpgradesPanel;
-import appeng.core.localization.GuiText;
-import appeng.menu.SlotSemantics;
-import com.glodblock.github.extendedae.container.ContainerWirelessExPAT;
-import net.minecraft.network.chat.Component;
-import net.minecraft.world.entity.player.Inventory;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class GuiWirelessExPAT extends GuiExPatternTerminal<ContainerWirelessExPAT> {
-
-    public GuiWirelessExPAT(ContainerWirelessExPAT menu, Inventory playerInventory, Component title, ScreenStyle style) {
-        super(menu, playerInventory, title, style);
-        this.widgets.add("upgrades", new UpgradesPanel(
-                menu.getSlots(SlotSemantics.UPGRADE),
-                this::getCompatibleUpgrades));
-        if (menu.getToolbox().isPresent()) {
-            this.widgets.add("toolbox", new ToolboxPanel(style, menu.getToolbox().getName()));
-        }
-    }
-
-    private List<Component> getCompatibleUpgrades() {
-        var list = new ArrayList<Component>();
-        list.add(GuiText.CompatibleUpgrades.text());
-        list.addAll(Upgrades.getTooltipLinesForMachine(menu.getUpgrades().getUpgradableItem()));
-        return list;
-    }
-
-}
