+++ src/main/resources/META-INF/neoforge.mods.toml
@@ -40,8 +40,14 @@ license="${mod_license}"
     ordering="AFTER"
     side="BOTH"
 [[dependencies.extendedae]]
-    modId="ae2wtlib"
-    type="OPTIONAL"
+    modId="ae2wtlib_api"
+    type="REQUIRED"
     versionRange="*"
     ordering="BEFORE"
+    side="BOTH"
+[[dependencies.extendedae]]
+    modId="ae2wtlib"
+    type="OPTIONAL"
+    versionRange="[19.1.0-alpha.1,)"
+    ordering="AFTER"
     side="BOTH"
