--- src/main/resources/assets/ae2/screens/u_ex_pat.json
@@ -1,23 +1,0 @@
-{
-  "$schema": "schema.json",
-  "includes": [
-    "ex_pattern_access_terminal.json",
-    "wtlib/universal_terminal.json",
-    "common/toolbox.json"
-  ],
-  "slots": {
-    "TOOLBOX": {
-      "bottom": 84,
-      "right": 20,
-      "grid": "BREAK_AFTER_3COLS"
-    }
-  },
-  "widgets": {
-    "toolbox": {
-      "right": 21,
-      "bottom": 90,
-      "width": 59,
-      "height": 66
-    }
-  }
-}
