+++ src/main/resources/assets/extendedae/lang/en_us.json
@@ -66,7 +66,6 @@
   "item.extendedae.threshold_export_bus": "ME Threshold Export Bus",
   "item.extendedae.oversize_interface_part": "ME Oversize Interface",
   "item.extendedae.entro_shard": "Entro Shard",
-  "item.ae2wtlib.wireless_ex_pattern_access_terminal": "Wireless Extended Pattern Access Terminal",
   "infinity.tooltip": "It is an infinity source.",
   "packaged_device.error.tooltip": "Invalid",
   "packaged_device.tooltip": "Device: %s",
